feat(ssr): style and script compilation and import
