+++ handler.go
@@ -48,6 +48,8 @@ func handleHome(w http.ResponseWriter, r *http.Request) {
 				Title:       "Architects Land",
 				Description: "Famille de SMP Minecraft privé",
 				Image:       "/static/terre-des-civilisations/background.webp",
+				Dark:        false,
+				Min:         false,
 			},
 			Seasons: seasonsData,
 		},
@@ -71,6 +73,8 @@ func handleRules(w http.ResponseWriter, r *http.Request) {
 				Title:       "Règles",
 				Description: "",
 				Image:       "/static/purgatory.webp",
+				Dark:        false,
+				Min:         true,
 			},
 		},
 	})
@@ -94,6 +98,8 @@ func handleTeam(w http.ResponseWriter, r *http.Request) {
 				Title:       "Règles",
 				Description: "",
 				Image:       "/static/purgatory.webp",
+				Dark:        false,
+				Min:         true,
 			},
 			Team: team,
 		},
@@ -130,6 +136,8 @@ func handleSeason(w http.ResponseWriter, r *http.Request) {
 				Title:       "Règles",
 				Description: "",
 				Image:       "/static/purgatory.webp",
+				Dark:        false,
+				Min:         true,
 			},
 			Season: season.toFullSeasonData(),
 		},
@@ -202,6 +210,8 @@ func (nf *NotFound) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 				Title:       "Perdu ?",
 				Description: "Il semblerait que vous vous êtes perdu·es dans le nether. Vous allez être redirigés dans l'overworld.",
 				Image:       "/static/nether.png",
+				Dark:        false,
+				Min:         false,
 			},
 		},
 	})
