--- resources/components/button/Button.tsx
@@ -1,14 +1,0 @@
-import "./Button.scss";
-
-export type Properties = {
-  href: string;
-  content: string;
-};
-
-export default function Button(props: Properties) {
-  return (
-    <a class={"btn"} href={props.href}>
-      {props.content}
-    </a>
-  );
-}
