--- resources/index.tsx
@@ -1,10 +1,0 @@
-/* @refresh reload */
-import { render } from "solid-js/web";
-import "./scss/main.scss";
-import App, { setupEvents } from "./App";
-
-const root = document.getElementById("root");
-
-render(() => <App />, root!);
-
-setupEvents();
