--- resources/pages/lost/Lost.tsx
@@ -1,15 +1,0 @@
-import Hero from "../../components/hero/Hero.tsx";
-
-export default function Lost() {
-  return (
-    <>
-      <Hero
-        image={"/nether.png"}
-        title={"Perdu ?"}
-        description={
-          "Il semblerait que vous vous êtes perdu·es dans le nether. Vous allez être redirigés dans l'overworld."
-        }
-      />
-    </>
-  );
-}
