--- resources/pages/rules/Rules.tsx
@@ -1,212 +1,0 @@
-import Hero from "../../components/hero/Hero.tsx";
-
-import "./Rules.scss";
-
-export default function Rules() {
-  return (
-    <>
-      <Hero image={"/purgatory.webp"} title={"Règles"} description={""} min />
-      <main>
-        <div class="content--large large-text tldr">
-          <h2>Résumé</h2>
-          <p>
-            La majorité des règles faisant appel au bon sens s'appliquent aux
-            saisons d'Architects Land.
-          </p>
-          <p>
-            Il est donc vivement conseillé d'être courtois·e, poli·e et
-            respectueux·se avec tou·tes. On évitera ainsi de parler de
-            politique, d'argent ou de religion. Chacun a sa sensibilité propre
-            qu'il est important de respecter. Si quelqu'un vous demande
-            d'arrêter, alors vous devez vous arrêter immédiatement.
-          </p>
-          <p>
-            Tout logiciel de triche est interdit. Toute exploitation de bug pour
-            obtenir un quelconque avantage est aussi prohibé.
-          </p>
-          <p>
-            Le mod <em>Simple Voice Chat</em> est obligatoire.
-          </p>
-        </div>
-        <div class="rules content--large large-text">
-          <div class="rules__item rules__attitude">
-            <h2>Attitude en jeu</h2>
-            <p>
-              Insulter, harceler ou discréditer quelqu'un est interdit, et l'est
-              dans tous les cas. Le racisme, le sexisme, l'homophobie, la
-              transphobie, la grossophobie, le validisme, l'islamophobie,
-              l'antisémitisme et autres sont donc évidemment interdits.
-            </p>
-            <p>
-              Les joueur·ses doivent avoir une attitude correcte en étant
-              courtois·e, poli·e et respectueux·se avec tou·tes. Les joueur·ses
-              sont alors encouragé·es de souhaiter la bienvenue aux nouveaux
-              joueur·ses, de remercier pour un service rendu et d'aider autrui.
-            </p>
-            <p>
-              Pour éviter de provoquer d'inutiles polémiques, les joueur·ses
-              doivent éviter de parler de la politique actuelle, qu'elle soit
-              mondiale, nationale ou locale, ainsi que d'argent ou de religion.
-              Architects Land n'a pas d'objectif militant.
-            </p>
-            <p>
-              Les joueur·ses doivent avoir conscience de la sensibilité d'autrui
-              et doivent la prendre en compte.
-            </p>
-            <p>
-              Les joueur·ses doivent respecter le consentement d'autrui, ou son
-              absence. Si un·e joueur·se demande d'arrêter, alors le·a ou les
-              joueur·ses doivent arrêter.
-            </p>
-          </div>
-          <div class="rules__item rules__gameplay">
-            <h2>Gameplay</h2>
-            <p>
-              Un modpack est fourni aux joueur·ses. Ils et elles sont libre de
-              ne pas l'utiliser.
-            </p>
-            <p>
-              L'unique mod obligatoire est <em>Simple Voice Chat</em>.
-            </p>
-            <p>
-              Toute modification du jeu donnant un avantage à un joueur est
-              interdite. Les logiciels de triches donnant accès à un X-Ray, une
-              Kill-Aura ou une Reach sont donc prohibés.
-            </p>
-            <p class={"has-subtitle"}>
-              Toute utilisation d'un bug donnant un avantage à un joueur est
-              interdite. Les bugs de duplication sont ainsi interdits. Les bugs
-              tolérés sont :
-            </p>
-            <ul>
-              <li>Cave-Ray du F5</li>
-              <li>Duplication de la TNT</li>
-              <li>Duplication des tapis</li>
-              <li>Duplication des fils</li>
-              <li>
-                Ticks d'invincibilité après un déco/reco si utilisé qu'une seule
-                fois
-              </li>
-            </ul>
-          </div>
-          <div class="rules__item rules__event">
-            <h2>Événements</h2>
-            <p>Les joueur·ses doivent indiquer leur présence aux événements.</p>
-            <p>
-              Si un joueur·se s'engage à être présent, alors il ou elle doit
-              être présent à l'événement. S'il ou elle ne peut pas, alors il
-              doit prévenir l'organisateur.
-            </p>
-            <p>
-              Chaque événement possède ses propres règles. Ces règles seront
-              communiquées avant le début de l'événement. Les règles
-              d'Architects Land sont toujours applicables.
-            </p>
-          </div>
-          <div class="rules__item rules__stream">
-            <h2>Enregistrement et streaming</h2>
-            <p>
-              En acceptant de rejoindre une saison d'Architects Land, les
-              joueur·ses ont aussi explicitement accepté que leur gameplay et
-              leur voix soient enregistrés pour être diffusés durant la ou les
-              saisons auxquelles le joueur participe par n'importe quel
-              joueur·se.
-            </p>
-            <p>
-              Tous les joueur·ses peuvent diffuser, en direct ou non, du contenu
-              provenant d'Architects Land tant que ce dernier n'enfreint pas les
-              règles d'Architects Land.
-            </p>
-            <p>
-              Si un·e joueur·se souhaite que sa voix ou que son gameplay ne soit
-              plus enregistrée dans l'objectif d'être diffusée, alors le·a
-              joueur·se doit quitter Architects Land.
-            </p>
-            <p>
-              Occasionnellement, un·e joueur·se peut informer un autre joueur·se
-              de sa volonté de ne pas être enregistré sur le moment. Cette
-              volonté doit être respecté.
-            </p>
-            <p>
-              Si un·e joueur·se d'Architects Land indique clairement et
-              distinctement que son contenu ne doit pas être regardé par les
-              autres joueur·ses, alors les joueur·ses doivent respecter sa
-              volonté.
-            </p>
-          </div>
-          <div class="rules__item rules__sanction">
-            <h2>Sanctions</h2>
-            <p>
-              Une sanction vise à rétablir le déséquilibre provoqué par un
-              manquement au règlement. Ce manquement ne provient pas
-              nécessairement d'un joueur.
-            </p>
-            <p>
-              L'unique personne habilitée à sanctionner un·e joueur·se est
-              l'organisateur de la saison. Cet organisateur est la même personne
-              possédant le site d'Architects Land. Actuellement, il s'agit de
-              William "Anhgelus Morhtuuzh" Hergès.
-            </p>
-            <p>
-              Les sanctions sont proportionnelles au déséquilibre provoqué. Si
-              un joueur utilise un bug interdit n'affectant que peu l'expérience
-              de jeu, alors il ne pourra pas encourir la sanction maximale.
-            </p>
-            <p>
-              Il n'existe que deux sanctions possibles : le blâme et le
-              bannissement.
-            </p>
-            <p>
-              Le blâme, ou warn, est un avertissement alertant le joueur de son
-              non respect des règles.
-            </p>
-            <p>
-              Le bannissement est le retrait du statue de joueur·se à un
-              joueur·se pour la saison. Le statue de joueur·se n'est pas
-              automatiquement accordé à la saison suivante : le joueur·se doit
-              en faire la demande qui peut être refusé.
-            </p>
-          </div>
-          <div class="rules__item rules__next_season">
-            <h2>Rejoindre Architects Land</h2>
-            <p>
-              A la fin d'une saison, les joueur·ses sont tous remerciés pour
-              leur participation.
-            </p>
-            <p>
-              Les joueur·ses ne sont pas automatiquement reconduits pour la
-              saison suivante. Ils doivent en faire la demande. Cette demande
-              peut être refusée. Cette demande est prioritaire par rapport aux
-              autres.
-            </p>
-            <p>
-              Si une personne souhaite rejoindre Architects Land, alors il ou
-              elle doit en faire la demande dans le temps imparti. Cette demande
-              peut être refusée.
-            </p>
-            <p>La demande doit répondre aux questions suivantes :</p>
-            <ul>
-              <li>Quel est votre pseudonyme Minecraft premium ?</li>
-              <li>
-                Pouvez-vous expliquer le principe d'Architects Land - Ruine des
-                Civilisations ?
-              </li>
-              <li>
-                Pourquoi souhaitez-vous rejoindre ce serveur et cette saison ?
-              </li>
-              <li>Comment avez-vous découvert ce serveur ?</li>
-            </ul>
-            <p>
-              La dernière question est facultative pour les joueur·ses des
-              saisons précédentes.
-            </p>
-            <p>
-              Généralement, une personne ne peut pas rejoindre une saison
-              d'Architects Land en cours.
-            </p>
-          </div>
-        </div>
-      </main>
-    </>
-  );
-}
