--- resources/pages/team/Team.scss
@@ -1,14 +1,0 @@
-@import "../../scss/variables";
-
-.team {
-  &__presentation {
-    background: $dark-background;
-    color: $light-sand;
-    text-align: center;
-
-    & p {
-      width: 75%;
-      margin: 0.5rem auto;
-    }
-  }
-}
