+++ resources/scss/global.scss
@@ -3,7 +3,7 @@
 @font-face {
   font-family: "Abril Fatface";
   font-weight: 400;
-  src: url("/fonts/Abril_Fatface/AbrilFatface-Regular.ttf") format("truetype");
+  src: url("/static/fonts/Abril_Fatface/AbrilFatface-Regular.ttf") format("truetype");
 }
 
 body {
