+++ resources/scss/main.scss
@@ -1,2 +1,11 @@
 @import "reset";
 @import "global";
+
+@import "specific/button";
+@import "specific/footer";
+@import "specific/hero";
+@import "specific/navbar";
+@import "specific/people";
+@import "specific/root";
+@import "specific/rules";
+@import "specific/season";
