+++ resources/scss/specific/hero.scss
