+++ resources/scss/specific/navbar.scss
