+++ resources/scss/specific/people.scss
