+++ resources/scss/specific/root.scss
